refactor: 알람 공통 infra 리팩터링
