--- src/main/java/com/example/sns/alarm/application/AlarmService.java
@@ -1,9 +1,0 @@
-package com.example.sns.alarm.application;
-
-import com.example.sns.alarm.infrastructure.dto.MessageDto;
-
-import java.util.List;
-
-public interface AlarmService {
-    void sendAll(List<MessageDto> messageDtoList);
-}
