+++ src/main/java/com/example/sns/common/infrastructure/fcm/AlarmService.java
@@ -1,0 +1,9 @@
+package com.example.sns.common.infrastructure.fcm;
+
+import com.example.sns.common.infrastructure.fcm.dto.MessageDto;
+
+import java.util.List;
+
+public interface AlarmService {
+    void sendAll(List<MessageDto> messageDtoList);
+}
