+++ src/main/java/com/example/sns/common/infrastructure/fcm/config/FcmConfig.java
@@ -1,4 +1,4 @@
-package com.example.sns.alarm.config;
+package com.example.sns.common.infrastructure.fcm.config;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
