+++ src/test/java/com/example/sns/alarm/application/AlarmEventHandlerTest.java
@@ -2,6 +2,7 @@ package com.example.sns.alarm.application;
 
 import com.example.sns.alarm.domain.Alarm;
 import com.example.sns.alarm.domain.AlarmRepository;
+import com.example.sns.common.infrastructure.fcm.AlarmService;
 import com.example.sns.common.support.ServiceTest;
 import com.example.sns.follow.application.FollowService;
 import com.example.sns.follow.application.FollowedEvent;
